% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel_requirement.R
\name{fuel_requirement}
\alias{fuel_requirement}
\title{Fuel Requirement}
\usage{
fuel_requirement(mass)
}
\arguments{
\item{mass}{A numeric vector}
}
\description{
Fuel required to launch a given module is based on its mass.
Specifically, to find the fuel required for a module,
take its mass, divide by three, round down, and subtract 2.
}
\examples{
sum(sapply(day_1, fuel_requirement))

}
