% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{get_solution}
\alias{get_solution}
\title{Get solution for a particular day}
\usage{
get_solution(day = 1)
}
\arguments{
\item{day}{integer between 1 and 25}
}
\description{
Get solution for a particular day
}
