% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{input_file}
\alias{input_file}
\title{Get path to input file}
\usage{
input_file(day)
}
\arguments{
\item{day}{integer between 1 and 25}
}
\value{
String, e.g. \verb{/path/to/input01.txt}
}
\description{
Get path to input file
}
