% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day04.R
\name{passwords_slow}
\alias{passwords_slow}
\title{Secure Container}
\usage{
passwords_slow(from = 273025, to = 767253)
}
\description{
You arrive at the Venus fuel depot only to discover it's protected by a password. The Elves had written the password on a sticky note, but someone threw it out.
}
\details{
However, they do remember a few key facts about the password:
\itemize{
\item It is a six-digit number.
\item The value is within the range given in your puzzle input.
\item Two adjacent digits are the same (like 22 in 122345).
\item Going from left to right, the digits never decrease; they only ever increase or stay the same (like 111123 or 135679).
}

Other than the range rule, the following are true:
\itemize{
\item 111111 meets these criteria (double 11, never decreases).
\item 223450 does not meet these criteria (decreasing pair of digits 50).
\item 123789 does not meet these criteria (no double).
}

How many different passwords within the range given in your puzzle input meet these criteria?
}
