% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_intersection.R
\name{port_intersection}
\alias{port_intersection}
\title{Port Intersection}
\usage{
port_intersection(wire1, wire2, use_time = FALSE)
}
\arguments{
\item{wire1, wire2}{A character vector of wires}

\item{use_time}{A logical should the answer return time or manhatan distance}
}
\description{
The wires twist and turn, but the two wires occasionally cross paths.
To fix the circuit, you need to find the intersection point closest to the central port.
Because the wires are on a grid, use the Manhattan distance for this measurement.
While the wires do technically cross right at the central port where they both start,
this point does not count, nor does a wire count as crossing with itself.
}
\examples{
port_intersection(day_3[[1]], day_3[[2]])
port_intersection(day_3[[1]], day_3[[2]], use_time = TRUE)

}
