% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-secure-container.R
\name{secure_container}
\alias{secure_container}
\title{Secure Container}
\usage{
secure_container(min, max = min, no_large_group = FALSE)
}
\arguments{
\item{min, max}{range of password}

\item{no_large_group}{}
}
\description{
Find password with following rules:
}
\details{
- It is a six-digit number.
- The value is within the range given in your puzzle input.
- Two adjacent digits are the same (like 22 in 122345).
- Going from left to right, the digits never decrease;
they only ever increase or stay the same (like 111123 or 135679).

If no_large_group is defined, the two adjacent matching digits are not part of
a larger group of matching digits.
}
\examples{
secure_container(245318, 765747)
secure_container(245318, 765747, no_large_group = TRUE)

}
